+++ R/enrich.family.R
@@ -148,7 +148,7 @@
     out
 }
 
-#' @method compute_theta family
+#' @export
 `compute_theta.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -182,7 +182,7 @@
     UseMethod('compute_theta')
 }
 
-#' @method compute_bfun family
+#' @export
 `compute_bfun.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -217,7 +217,7 @@
 }
 
 
-#' @method compute_c1fun family
+#' @export
 `compute_c1fun.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -252,7 +252,7 @@
 }
 
 
-#' @method compute_c2fun family
+#' @export
 `compute_c2fun.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -286,7 +286,7 @@
     UseMethod('compute_c2fun')
 }
 
-#' @method compute_d1variance family
+#' @export
 `compute_d1variance.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -340,7 +340,7 @@
 }
 
 
-#' @method compute_d2variance family
+#' @export
 `compute_d2variance.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -393,7 +393,7 @@
     UseMethod('compute_d2variance')
 }
 
-#' @method compute_afun family
+#' @export
 `compute_afun.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -429,7 +429,7 @@
 }
 
 
-#' @method compute_d1afun family
+#' @export
 `compute_d1afun.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -466,7 +466,7 @@
 }
 
 
-#' @method compute_d2afun family
+#' @export
 `compute_d2afun.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -501,7 +501,7 @@
     UseMethod('compute_d2afun')
 }
 
-#' @method compute_d3afun family
+#' @export
 `compute_d3afun.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -536,7 +536,7 @@
 }
 
 
-#' @method compute_d4afun family
+#' @export
 `compute_d4afun.family` <- function(object, ...) {
     family <- object$family
     switch(family,
@@ -581,6 +581,5 @@
 ##     component = list("d1variance", "d2variance", "d1afun", "d2afun",
 ##         "d3afun", c("d1variance", "d2variance"), c("d1afun",
 ##             "d2afun", "d3afun")), path = "~/Downloads", attempt_rename = TRUE)
-